requirements.txt
numpy
scipy
matplotlib
pillow
pywt
torch
torchvision
scikit-learn
tqdm